+++ RELEASE-NOTES.md
@@ -38,9 +38,11 @@
 * Added support for Mac Os x86_64 architecture.
 * Added image repository model in snowflake.yml.
 * Added `snow spcs service deploy` command.
-* Added notebooks to `snow object` commands.
-* Added `snow notebook deploy` command that allows creating a notebook using local file.
-* Add support for containerized notebooks.
+* Changes to notebooks commands:
+  * Added `snow notebook deploy` command that allows creating a notebook using local file.
+  * Added support for snowflake.yml for `snow notebook create` command.
+  * Add support for [containerized notebooks](https://docs.snowflake.com/en/user-guide/ui-snowsight/notebooks-on-spcs).
+  * Added notebooks to `snow object` commands.
 * Added `--force-replace` flag in `snow snowpark deploy` command to replace entities even if no changes are detected.
 
 ## Fixes and improvements